requirements.txt
slack-sdk>=3.21.3
python-dotenv>=1.0.1
slack-bolt>=1.20.0
slack-logger>=0.1.0
pytest>=pytest-8.3.5 

##pip install -r requirements.txt
